webapp/requirements.txt
django==4.2.*
gunicorn
python-decouple
lightrag-hku[ollama]
